Move type definitions to specfile_finder module
Moved ElementType and DocumentType from loader to specfile_finder:
- These types are file path resolvers, not test harness concerns
- Added convenience helpers in specfile_finder:
  - find_element_file(ElementType, number)
  - find_document_file(DocumentType, number)
  - list_element_numbers(ElementType)

Updated loader.rs to re-export types and delegate to helpers:
- Removed ~50 lines of type definitions
- Re-exports types for backward compatibility
- Delegates to specfile_finder convenience functions

Benefits:
- specfile_finder is now self-contained (types + functions)
- loader.rs is simpler (just test harness API)
- Clear module boundaries

All 385 tests pass.
